Update Info.plist metadata with CossackLabs branding
Change the bundle ID to something related to us so that we're not
infringing on "org.openssl" namespace (used by the original project),
and have the binary we build clearly associated with us.

Similarly, we are going to offer OpenSSL 1.0.2u, hence update the
marketing version of the bundle to reflect that. Also update the
default version in the build script so that we can use it directly.