Apple Silicon support
Add architecture configuration for OpenSSL based on changes suggested
by Sasmito [1]. Since we're interested only in macOS and iOS, add only
those simulators, cross-compiled for arm64.

This only lets the compiler to compile something as support for arm64
is not complete in OpenSSL [2]. For example, you can see that we disable
inline assembly, so the code might actually run slower. It's not much of
an issue right now though since there is no arm64 hardware yet on the
public market.

[1]: https://github.com/balthisar/openssl-xcframeworks/pull/6
[2]: https://github.com/openssl/openssl/issues/12254

Suggested-by: Sasmito Adibowo <adib@basil-salad.com>