Restore removed iOS 10.0 targets
Upstream has moved to using iOS 12.0 as the default minimum supported
version. With this, several architectures has been removed in commit
2be1214e3b1f87478e531c82056a4d9f3c895dc2. However, we are still using
iOS 10.0 as our minimum supported version and still the removed 32-bit
architectures: armv7 and arm7s for iOS devices and i386 for simulators.
If they are not included in the binary frameworks, Xcode will complain
about missing architecture slices.