Adjust packaging to XCFramework
Upstream has moved to packaging a single XCFramework with all targets in
it. This brings good news as now Apple Silicon will be supported for iOS
Simulators, removing the need for some hacks. This also brings bad news
because this is a breaking change for CLOpenSSL users via Carthage.

While CocoaPods will automatically figure out how to link in the
vendored XCFramework into the application, Carthage users will have to
adjust their binary dependencies to download the unified XCFramework
instead of two separate frameworks for iOS and macOS. They will also
need to adjust their Xcode projects to use "openssl.xcframework" instead
of "openssl.framework". Nothing we can do about it.

This is also a breaking change for our build system as the upstream
build system now builds iOS (device) and iOS Simulator frameworks
separately. This means that there are two "openssl.framework" binaries
for iOS and neither is usable with both iOS and iOS Simulator, as it was
before. We can't package them as an "iPhone" target to maintain backward
compatibility with previous Carthage packaging.

Overall, this means that the next release of CLOpenSSL is a flag day for
its users. Users of the previous version MUST either migrate their builds
immediately, or pin the previous version 1.1.10801 until they migrate.